Add .install

+++ backup.sh
@@ -7,20 +7,24 @@ echo "Doing a backup of $DEVICE."
 
 TARGET=/var/lib/flasher/backups/
 
-# Attempt 1: find a mountable partition, mount it and backup the content.
+FILENAME=$TARGET/$(date +%Y-%m-%d_%H:%M)
+
+
 PARTITION=${DEVICE}1
-MOUNT=$(mktemp -d)
-mount $PARTITION $MOUNT
+if [ -b $PARTITION ]
+then
+	# Attempt 1: find a mountable partition, mount it and backup the content.
+	MOUNT=$(mktemp -d)
+	mount $PARTITION $MOUNT
 
-FILENAME=$(date +%Y-%m-%d_%H:%M)
-tar -C $MOUNT -cf - . | zstd > $TARGET/$FILENAME.tar.zst
-ls -l $MOUNT/ >> /var/log/flasher.log
+	tar -C $MOUNT -cf - . | zstd > $FILENAME.tar.zst
+	ls -l $MOUNT/ >> /var/log/flasher.log
 
-umount $PARTITION
-rmdir $MOUNT
+	umount $PARTITION
+	rmdir $MOUNT
+else
+	# Attempt 2: just backup the block device.
+	echo Meh~
+	cat $DEVICE | zstd > $FILENAME.img.zst
+fi
 
-exit 0
-
-
-# Attempt 2: just backup the block device.
-echo Meh~
